add code for recieving packets
git-svn-id: file:///home/lennart/svn/public/avahi/trunk@11 941a03a8-eaeb-0310-b9a0-b1bbd8fe43fe

+++ socket.h
@@ -11,7 +11,7 @@ gint flx_open_socket_ipv6(void);
 gint flx_send_dns_packet_ipv4(gint fd, gint iface, flxDnsPacket *p);
 gint flx_send_dns_packet_ipv6(gint fd, gint iface, flxDnsPacket *p);
 
-flxDnsPacket *flx_recv_dns_packet_ipv4(gint fd, struct sockaddr_in*ret_sa, gint *ret_iface, gint *ret_ttl);
-flxDnsPacket *flx_recv_dns_packet_ipv6(gint fd, struct sockaddr_in6*ret_sa, gint *ret_iface, gint *ret_ttl);
+flxDnsPacket *flx_recv_dns_packet_ipv4(gint fd, struct sockaddr_in*ret_sa, gint *ret_iface, guint8 *ret_ttl);
+flxDnsPacket *flx_recv_dns_packet_ipv6(gint fd, struct sockaddr_in6*ret_sa, gint *ret_iface, guint8 *ret_ttl);
 
 #endif
